article-to-speech/requirements.txt
gTTS
google-genai
beautifulsoup4
requests
streamlit
pydub
python-dotenv
firebase-admin
google-cloud-storage
watchdog
google-cloud-texttospeech>=2.31.0
fastapi
uvicorn
python-multipart
